+++ src/app.module.ts
@@ -2,11 +2,11 @@ import { Module, OnApplicationBootstrap, Logger } from '@nestjs/common';
 import { DatabaseModule } from './database/database.module';
 import { AuthGuard } from 'src/middlewares/auth-module/auth';
 import { UsuarioModule } from './modules/usuario/usuario.module';
-import { ProdutosModule } from './modules/produtos/produtos.module';
-import { EntradasModule } from './modules/entradas/entradas.module';
+import { ProdutosModule } from './modules/produto/produtos.module';
+import { EntradasModule } from './modules/entrada/entradas.module';
 import { AuthModule } from './middlewares/auth-module/auth-module.module';
 import { FornecedorModule } from './modules/fornecedor/fornecedor.module';
-import { GrupoModule } from './modules/grupos/grupo.module';
+import { GrupoModule } from './modules/grupo/grupo.module';
 
 
 @Module({
