+++ src/database/database.config.ts
@@ -2,9 +2,9 @@ import { Sequelize } from 'sequelize-typescript';
 import pg from 'pg';
 import { Usuario } from 'src/modules/usuario/usuario.model';
 import { Fornecedor } from 'src/modules/fornecedor/fornecedor.model';
-import { Entrada } from 'src/modules/entradas/entradas.model';
-import { Grupo } from 'src/modules/grupos/grupo.model';
-import { Produto } from 'src/modules/produtos/produto.model';
+import { Entrada } from 'src/modules/entrada/entradas.model';
+import { Grupo } from 'src/modules/grupo/grupo.model';
+import { Produto } from 'src/modules/produto/produto.model';
 
 //Database config
 export const databaseProviders = [
