+++ src/modules/entrada/entradas.model.ts
