+++ src/modules/entrada/use-cases/create-entrada/create-entrada.controller.ts
