+++ src/modules/entrada/use-cases/create-entrada/create-entrada.dto.ts
