+++ src/modules/entrada/use-cases/delete-entrada/delete-entrada.controller.ts
