+++ src/modules/entrada/use-cases/delete-entrada/delete-entrada.service.ts
