+++ src/modules/entrada/use-cases/get-entrada/get-entrada.service.ts
