+++ src/modules/entrada/use-cases/update-entrada/update-entrada.controller.ts
