+++ src/modules/entrada/use-cases/update-entrada/update-entrada.service.ts
