+++ src/modules/grupo/use-cases/create-grupo/create-grupo.controller.ts
