+++ src/modules/grupo/use-cases/delete-grupo/delete-grupo.controller.ts
