+++ src/modules/grupo/use-cases/delete-grupo/delete-grupo.service.ts
