+++ src/modules/grupo/use-cases/get-grupo/get-grupo.controller.ts
